separated utility bar into its own component
